build dispatcher on first call to getDispatcher.
this adds a class variable to RouteCollection which holds a Dispatcher. The dispatcher
is built on the first call to getDispatcher, and returned. Subsequent calls to
getDispatcher return the same instance.
Further addition of routes is disallowed after this point, and an Exception is
thrown if you try to add a route after building the dispatcher.