Fix route-specific strategies
This commit fixes a bug where calling `setStrategy()` on a route had no
effect. The route-specific strategy would always be replaced by the
group strategy (if the route was in a group) or the global Router
strategy (if the route was not in a group). With this commit, the
strategy on the route takes precedence over group and global strategies,
as intended.