Ensure League\Route\Dispatcher::dispatchRequest handlers are League\Route\Route
Background:

There seems to be an implicit contract in `League\Route\Dispatcher::dispatchRequest`
that match[1] (the handler) is a `League\Route\Route`
(as it calls [setVars](https://github.com/thephpleague/route/blob/4.2.0/src/Dispatcher.php#L42))

This throws when the route is created via `Router::addRoute`, because now $match[1] will be a callable|mixed.

A possible solution is included in this PR. Ensure that we always get a
Router.